Completely refactor gitlab_ci_runner::runner
Before this, the complete hash with all runners was injected into
gitlab_ci_runner::runner. Based on the title of the resource it was then
decided which runner should be extracted from the global hash and merge
with defaults ect.

This was completely refactored in a way that a single
gitlab_ci_runner::runner resource now only holds all data which is need
to configure it but no other.

This also changes the default of gitlab_ci_runner::runners(_defaults) to
an empty hash because I don't see why we need to force user to configure
it.