Make manage_docker optional and include the docker module if used
We need to include the complete docker module to ensure docker is
installed on the host to be able to use `docker::images`.

Docker is by far not a required dependency for a Gitlab Runner instance
and should therefore not be installed/configured by default. This also
makes everything easier because we still support CentOS 6 and
puppetlabs-docker doesn't.